privcall: Add ability to switch boxes
The privcall table now includes a function to switch boxes. This can be
used, for example, by RTOS to switch processes.

The commit also adds a pointer to the `g_active_box` which is needed to
get information on the currently active box.

Handlers for thread-specific functions have been removed, as not needed
any more.