SVC: Remove the metadata from the macro
At the moment no macro appends metadata explicitly to an SVCall. We
don't have secure gateways any more and register gateways append
metadata to a regular memory access (in a different way anyway, i.e.,
using a C struct), so it's cleaner to have the `UVISOR_SVC(...)` macro
not use the metadata field.

If needed, the `UVISOR_SVC_METADATA(...)` provides the old API.