vMPU: Use centralized structure for SRAM sections
This commit helps towards the centralization of the SRAM configuration
for every box. Every structure that must be allocated by the uVisor box
enumeration function is now grouped into a UvisorBssSections.

This is a purely semantic change. The actual functions for box
enumeration will be refactored in a later commit.